requirements.txt
bigtree==0.9.2
Flask==2.3.2
Flask_Migrate==4.0.4
flask_sqlalchemy==3.0.3
langchain==0.0.179
openai==0.27.6
PyGithub==1.58.2
python-dotenv==1.0.0
redis==4.5.5
rq==1.14.1
psycopg2-binary==2.9.6